requirements.txt
elevenlabs
fastapi
jinja2
python-dotenv
uvicorn